A "weird" solution to the convert challange
This solution is not a standard way to write the methods to solve our
problem, but it is not something that you never see, either.  Usually
when someone does this, the write those methods in a programatic way,
rather than taking the time to type them out.  This makes it simple to
make thousands, hundreds, or even tens of methods that are similar yet
different.

I did it the hard way!